Move code in physics::types module to physics
