+++ src/util/gens.rs
@@ -8,7 +8,7 @@ use rand::ThreadRng;
 use crate::geometry::types::Point;
 use crate::geometry::types::Vector;
 use crate::geometry::util::Transformation;
-use crate::physics::types::Mass;
+use crate::physics::Mass;
 
 // Generator /////////////////////////////////////////////////////////////////
 //
@@ -229,7 +229,7 @@ mod tests {
 
     use crate::geometry::types::Point;
     use crate::geometry::types::Vector;
-    use crate::physics::types::Mass;
+    use crate::physics::Mass;
     use crate::util::gens::*;
 
     #[test]
